Create README.md
The content of this repository was created for presentation purposes onl and it has not been fully tested..  The code does not include all possible aspects of responsive web design and does not guarantee it will work in your project.  Do not use this code in a production environment until you have fully test it.